Make function context aware
If there is an expression that does not end with operation appending a
function does to produce a valid result.

i.e. if we have typed 42 then type sqrt( appending would produce 42sqrt(
which is not a valid expression.

This commit changes the logic so that the function inserted at the start
of the buffer when the resulting expression wouldn't be correct, i.e.
pressing sqrt( when 42 was in the buffer would produce sqrt(42 instead.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>